+++ src/BookFx/Epplus/Constraint.cs
@@ -16,13 +16,13 @@
 
         internal const int MaxIndentSize = 250;
 
-        internal const float MinColSize = 0f;
+        internal const double MinColSize = 0d;
 
-        internal const float MaxColSize = 255f;
+        internal const double MaxColSize = 255d;
 
-        internal const float MinRowSize = 0f;
+        internal const double MinRowSize = 0d;
 
-        internal const float MaxRowSize = 409f;
+        internal const double MaxRowSize = 409d;
 
         internal const double MinMargin = 0;
